Extract common types into separate module
